% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{bread.hful}
\alias{bread.hful}
\title{Calculate the bread for HFUL asymptotic var-cov matrix}
\usage{
bread.hful(X, Z, alpha.hat, projection.matrix.z = NULL)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{Z}{An nxp matrix of regressors from the first-stage equation. The included instruments (from the structural equation) must be listed first}

\item{alpha.hat}{\eqn{\hat\alpha} used when computing the estimator}

\item{projection.matrix.z}{An optional argument to input the projection matrix of Z if it is already computed.}
}
\description{
Calculates \eqn{\hat H=X'P X-\sum_i X_i P_{ii} X_i'-\hat\alpha X'X}, the bread matrix for HFUL estimators.
}
