% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{bread.jive}
\alias{bread.jive}
\title{Calculate the bread for JIVE asymptotic var-cov matrix}
\usage{
bread.jive(X, Z, projection.matrix.z = NULL)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{Z}{An nxp matrix of regressors from the first-stage equation. The included instruments (from the structural equation) must be listed first}

\item{projection.matrix.z}{An optional argument to input the projection matrix for Z if it is already computed.}
}
\description{
Calculates \eqn{\hat H=X'PX-\sum_i X_i P_{ii} X_i'}, the bread matrix for JIVE estimators.
}
