% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{bread.kclassliml}
\alias{bread.kclassliml}
\title{Calculate the bread for LIML asymptotic var-cov matrix}
\usage{
bread.kclassliml(
  X,
  Z,
  n.endogenous,
  k,
  elimination.matrix.z = NULL,
  x.tilde = NULL
)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{Z}{An nxp matrix of regressors from the first-stage equation. The included instruments (from the structural equation) must be listed first}

\item{n.endogenous}{The number of endogenous variables in X.}

\item{k}{A constant to be used for the k-class estimator. Must always be provided but has no effect if x.tilde is provided.}

\item{elimination.matrix.z}{An optional argument to input the elimination matrix for Z if it is already computed.}

\item{x.tilde}{An optional argument to input x.tilde if it is already computed. If x.tilde is input, elimination.matrix.z and k has no effect. If x.tilde is not input, k is required and the computation is accelerated by inputting elimination.matrix.z}

\item{type}{The type of var-cov matrix to calculate. Defaults to HC1. Currently supports HC0 & HC1}
}
\description{
Calculates \eqn{\tilde{X}'X}, the bread matrix for LIML.
}
