% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{bread.rjive}
\alias{bread.rjive}
\title{Calculate the bread for RJIVE asymptotic var-cov matrix (todo)}
\usage{
bread.rjive(X, Z, projection.matrix.z = NULL, penalty.matrix.z = NULL)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{Z}{An nxp matrix of regressors from the first-stage equation. The included instruments (from the structural equation) must be listed first}

\item{projection.matrix.z}{An optional argument to input the augmented projection matrix $Z(Z'Z+\Lambda'\Lambda)^{-1}Z'$ for Z if it is already computed.}

\item{penalty.matrix.z}{An optional argument to input the penalty matrix $\Lambda$ for Z if it is already computed.}
}
\description{
Calculates \eqn{\hat H=X'P^A X-\sum_i X_i P_{ii}^A X_i'}, the bread matrix for RJIVE estimators.
}
