% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{bread.tsls}
\alias{bread.tsls}
\title{Calculate the bread for 2SLS asymptotic var-cov matrix}
\usage{
bread.tsls(X, Z, var.z = NULL, cov.xz = NULL)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation}

\item{Z}{An nxp matrix of regressors from the first-stage equation}

\item{var.z}{Optional input. $n$ times variance-covariance matrix of Z ($Z'Z$)}

\item{cov.xz}{Optional input. $n$ times the covariance between X and Z ($X'Z$)}
}
\description{
Calculates \eqn{X'Z(Z'Z)^{-1}Z'X}, the bread matrix for 2SLS
}
