% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-computation-functions.R
\name{compute_efficient_weighting}
\alias{compute_efficient_weighting}
\title{Compute efficient weighting matrix for GMM.}
\usage{
compute_efficient_weighting(Z, residuals)
}
\arguments{
\item{Z}{An nxp matrix of regressors from the first-stage equation. The included instruments (from the structural equation) must be listed first}

\item{residuals}{An nx1 vector of residuals from the first GMM estimation.}
}
\description{
Calculates the efficient weighting matrix for the twostep linear GMM estimator.
}
