% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-computation-functions.R
\name{compute_hful_alpha.tilde}
\alias{compute_hful_alpha.tilde}
\title{Compute alpha corresponding to HFUL in the general HFUL estimator}
\usage{
compute_hful_alpha.tilde(
  y,
  X,
  Z,
  C = 1,
  elimination.matrix.z = NULL,
  leverages = NULL,
  weighted.x.bar = NULL,
  x.bar.hat = NULL
)
}
\arguments{
\item{y}{An nx1 vector of outcomes}

\item{X}{An nxk matrix of regressors from the structural equation.}

\item{Z}{An nxp matrix of regressors from the first-stage equation.}

\item{C}{A constant. Default is 1, which is recommended by Hausman et al.}

\item{elimination.matrix.z}{An optional argument to input the elimination matrix for Z if it is already computed.}

\item{leverages}{An optional argument to input the leverages if they are already computed.}

\item{weighted.x.bar}{An optional argument to input the $\sum_{i=1}^n P_{ii}\bar{X}_i\bar{X}_i'=X'diag(P)X$ if it is already computed.}

\item{x.bar.hat}{An optional argument to input $\bar{X}P$=[y, X]\;P$ if it is already computed}
}
\description{
Calculates the alpha hat suggested by Hausman et al (2012).
}
