% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-computation-functions.R
\name{compute_rjive_penalty}
\alias{compute_rjive_penalty}
\title{Compute penalty of RJIVE}
\usage{
compute_rjive_penalty(X, n.endogenous, n.instruments)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{n.endogenous}{The number of endogenous variables in X.}

\item{n.instruments}{The number of instruments in Z.}
}
\description{
Calculates the penalty following the formula suggested by Hansen & Kozbur (2014).
}
