% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endo-test.R
\name{endo.SHat}
\alias{endo.SHat}
\title{Relevant Instrumental Variable Selection for endogeneity test}
\usage{
endo.SHat(
  ITT_Y,
  ITT_D,
  WUMat,
  SigmaSqY,
  SigmaSqD,
  SigmaYD,
  covW,
  bootstrap = FALSE,
  tuning = 2.01
)
}
\arguments{
\item{ITT_Y}{a p_z by 1 numeric vector denoting the estimated coefficients of instruments in the treatment model.}

\item{ITT_D}{a p_z by 1 numeric vector denoting the estimated coefficients of instruments in the treatment model.}

\item{WUMat}{a numeric matrix denoting WU where U is the precision matrix of W and W is the instrument-covariate matrix (Z, X).}

\item{SigmaSqY}{a numeric scalar denoting the consistent estimator of the noise level in the outcome model.}

\item{SigmaSqD}{a numeric scalar denoting the consistent estimator of the noise level in the treatment model.}

\item{SigmaYD}{a numeric scalar denoting the consistent estimator of the covariance between the error term in the treatment model and the error term in the outcome model.}

\item{covW}{a numeric, non-missing matrix that computes the sample covariance of W}

\item{bootstrap}{a logical value, default by FALSE.}

\item{tuning}{a numeric scalar value tuning parameter for TSHT greater 2, with default 2.01.}
}
\value{
\item{\code{SHat}}{a numeric vector denoting the set of relevant IVs.}
}
\description{
Implementation of Hard Thresholding for relevant instrumental variable selection. This function only takes the necessary inputs of the functions \code{\link{TSHT.OLS}} or \code{\link{TSHT.DeLasso}}. Any other methods to compute the necessary inputs can be adopted by the users according to their preferences.
}
