% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endo-test.R
\name{endo.test}
\alias{endo.test}
\title{Endogeneity-test}
\usage{
endo.test(
  Y,
  D,
  Z,
  X,
  intercept = TRUE,
  alpha = 0.05,
  method = "DeLasso",
  invalid = TRUE,
  voting = "MaxClique"
)
}
\arguments{
\item{Y}{A continuous vector of outcomes.}

\item{D}{A continuousvector of endogenous variables.}

\item{Z}{A matrix of instruments.}

\item{X}{A matrix of exogenous covariates.}

\item{intercept}{Should the intercept be included? Default is \code{TRUE} and if so, you do not need to add a column of 1s in X.}

\item{alpha}{The significance level for the confidence interval. (default = 0.05)}

\item{method}{The method which will be used to estimate the reduced form parameters in TSHT. "OLS" stands for ordinary least square and "DeLasso" stands for the debiased Lasso estimator. (default = "DeLasso")}

\item{invalid}{If \code{TRUE}, the method is robust to the presence of possibly invalid IVs; If \code{FALSE}, the method assumes all IVs to be valid. (default = TRUE)}

\item{voting}{The voting option used to estimate valid IVs. 'MP' stnads for majority and plurality voting, 'MaxClique' stands for maximum clique in the IV voting matrix. (default = 'MaxClique')}
}
\value{
\code{endo.test} returns an object of class "endotest".
    An object class "endotest" is a list containing the following components:
   \item{\code{Q}}{Endogeneity test statistic.}
   \item{\code{Sigma12}}{Estimated covaraince.}
   \item{\code{VHat}}{The estimated set of relevant and vaild IVs.}
   \item{\code{alpha}}{The significance level.}
}
\description{
Conduct the endogeneity test in high-dimensional linear models in the presence of potentially invalid instrumental variables.
}
\examples{
\dontrun{
n = 500; L = 600; s = 3; k = 10; px = 10;
alpha = c(rep(3,s),rep(0,L-s)); beta = 1; gamma = c(rep(1,k),rep(0,L-k))
phi<-(1/px)*seq(1,px)+0.5; psi<-(1/px)*seq(1,px)+1
epsilonSigma = matrix(c(1,0.8,0.8,1),2,2)
Z = matrix(rnorm(n*L),n,L)
X = matrix(rnorm(n*px),n,px)
epsilon = MASS::mvrnorm(n,rep(0,2),epsilonSigma)
D =  0.5 + Z \%*\% gamma + X \%*\% psi + epsilon[,1]
Y = -0.5 + Z \%*\% alpha + D * beta + X \%*\% phi + epsilon[,2]
TSHT(Y,D,Z,X, method = "DeLasso")
endo.test.model <- endo.test(Y,D,Z,X)
summary(endo.test.model)
}



}
