% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-functions.R
\name{gmm.fit}
\alias{gmm.fit}
\title{Fit GMM moments estimator given weighting matrix W}
\usage{
gmm.fit(y, X, Z, W = NULL)
}
\arguments{
\item{y}{An nx1 vector of outcomes}

\item{X}{An nxk matrix of regressors from the structural equation}

\item{Z}{An nxp matrix of regressors from the first-stage equation}

\item{W}{A p.s.d. weighting matrix. If no argument is supplied, (Z'Z)^-1 is used corresponding to 2SLS.}
}
\description{
Fits the linear GMM estimator given the outcomes, instruments, structural regressors and a weighting matrix.
}
