% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-functions.R
\name{hful.fit}
\alias{hful.fit}
\title{HFUL formula given alpha hat in the general HFUL estimator}
\usage{
hful.fit(
  y,
  X,
  Z,
  alpha.hat,
  elimination.matrix.z = NULL,
  leverages = NULL,
  weighted.x.bar = NULL,
  x.bar.hat = NULL
)
}
\arguments{
\item{y}{An nx1 vector of outcomes}

\item{X}{An nxk matrix of regressors from the structural equation.}

\item{Z}{An nxp matrix of regressors from the first-stage equation.}

\item{alpha.hat}{The alpha hat computed according to the HFUL procedure.}

\item{elimination.matrix.z}{An optional argument to input the elimination matrix for Z if it is already computed.}

\item{leverages}{An optional argument to input the leverages if they are already computed.}

\item{weighted.x.bar}{An optional argument to input the $\sum_{i=1}^n P_{ii}\bar{X}_i\bar{X}_i'=X'diag(P)X$ if it is already computed.}

\item{x.bar.hat}{An optional argument to input $\bar{X}P$=[y, X]\;P$ if it is already computed}
}
\description{
Computes  suggested by Hausman et al (2012).
}
