% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov.R
\name{hful.vcov}
\alias{hful.vcov}
\title{Calculate HFUL asymptotic var-cov matrix}
\usage{
hful.vcov(X, Z, e.hat, alpha.hat, projection.matrix.z = NULL, SSR = NULL)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{Z}{An nxp matrix of regressors from the first-stage equation. The included instruments (from the structural equation) must be listed first}

\item{e.hat}{An nx1 vector of residuals from the structural equation}

\item{alpha.hat}{The alpha hat computed according to the HFUL procedure.}

\item{projection.matrix.z}{An optional argument to input the elimination matrix for Z if it is already computed.}

\item{SSR}{Optional. The sum of squared residuals from the structural equation.}
}
\description{
Calculates the HFUL var-cov matrix given asymptotic theory. The HFUL variance-covariance matrix is derived in \insertCite{hausman2012hful}{RobustIV}.
}
\references{
\insertRef{hausman2012hful}{RobustIV}
}
