% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov.R
\name{jive.vcov}
\alias{jive.vcov}
\title{Calculate JIVE asymptotic var-cov matrix}
\usage{
jive.vcov(X, Z, e.hat)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation}

\item{Z}{An nxp matrix of regressors from the first-stage equation}

\item{e.hat}{An nx1 vector of residuals from the structural equation}
}
\description{
Calculates the JIVE var-cov matrix given asymptotic theory
}
