% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-functions.R
\name{jive1.fit}
\alias{jive1.fit}
\title{Fit JIVE1 estimator}
\usage{
jive1.fit(y, X, Z)
}
\arguments{
\item{y}{An nx1 vector of outcomes}

\item{X}{An nxk matrix of regressors from the structural equation}

\item{Z}{An nxp matrix of regressors from the first-stage equation}
}
\description{
Fits the JIVE1 estimator given the outcomes, instruments and structural regressors.
}
