% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-functions.R
\name{kclass.fit}
\alias{kclass.fit}
\title{Fit k-classs estimator given k}
\usage{
kclass.fit(y, X, Z, k, elimination.matrix.z = NULL)
}
\arguments{
\item{y}{An nx1 vector of outcomes}

\item{X}{An nxk matrix of regressors from the structural equation}

\item{Z}{An nxp matrix of regressors from the first-stage equation}

\item{k}{A constant to be used for the k-class estimator}

\item{elimination.matrix.z}{An optional argument to input the elimination matrix for Z if it is already computed.}
}
\description{
Fits the k-class estimator given a formula and a dataset. Given an appropriate choice of k, this function also fits the LIML estimator.
}
