% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{meat.hful}
\alias{meat.hful}
\title{Calculate the meat for HFUL asymptotic var-cov matrix}
\usage{
meat.hful(X, Z, e.hat, alpha.hat, projection.matrix.z = NULL, SSR)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{Z}{An nxp matrix of regressors from the first-stage equation. The included instruments (from the structural equation) must be listed first}

\item{e.hat}{An nx1 vector of residuals from the structural equation}

\item{alpha.hat}{\eqn{\hat\alpha} used when computing the estimator}

\item{projection.matrix.z}{An optional argument to input the projection matrix of Z if it is already computed.}

\item{SSR}{Optional. The sum of squared residuals from the structural equation}
}
\description{
Calculates the meat matrix for HFUL estimators.
}
