% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{meat.jive}
\alias{meat.jive}
\title{Calculate the meat for JIVE asymptotic var-cov matrix}
\usage{
meat.jive(X, Z, e.hat, projection.matrix.z = NULL)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation.}

\item{Z}{An nxp matrix of regressors from the first-stage equation.}

\item{e.hat}{An nx1 vector of residuals from the structural equation}

\item{projection.matrix.z}{An optional argument to input the elimination matrix for Z if it is already computed.}
}
\description{
Calculates \eqn{\hat{\Sigma}'X}, the bread matrix for JIVE estimators.
}
