% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{meat.kclassliml}
\alias{meat.kclassliml}
\title{Calculate the meat for LIML asymptotic var-cov matrix}
\usage{
meat.kclassliml(
  X,
  Z,
  e.hat,
  n.endogenous,
  k,
  elimination.matrix.z = NULL,
  x.tilde = NULL,
  type = "hc1"
)
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{Z}{An nxp matrix of regressors from the first-stage equation. The included instruments (from the structural equation) must be listed first}

\item{e.hat}{An nx1 vector of residuals from the structural equation}

\item{n.endogenous}{The number of endogenous variables in X.}

\item{k}{A constant to be used for the k-class estimator}

\item{elimination.matrix.z}{An optional argument to input the elimination matrix for Z if it is already computed.}

\item{x.tilde}{An optional argument to input x.tilde if it is already computed. If x.tilde is input, elimination.matrix.z and k has no effect.}

\item{type}{The type of var-cov matrix to calculate. Defaults to HC1. Currently supports HC0 & HC1}
}
\description{
Calculates \eqn{\tilde{X}'X}, the bread matrix for LIML.
}
