% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov-helpers.R
\name{meat.tsls}
\alias{meat.tsls}
\title{Calculate the meat for 2SLS asymptotic var-cov matrix}
\usage{
meat.tsls(X, Z, e.hat, var.z = NULL, cov.xz = NULL, type = "hc1")
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation}

\item{Z}{An nxp matrix of regressors from the first-stage equation}

\item{e.hat}{An nx1 vector of residuals from the structural equation}

\item{var.z}{Optional input. $n$ times the variance-covariance matrix of Z ($Z'Z$)}

\item{cov.xz}{Optional input. $n$ times the covariance between X and Z ($X'Z$)}

\item{type}{The type of var-cov matrix to calculate. Defaults to HC1. Currently supports HC0 & HC1}
}
\description{
Calculates \eqn{X'Z(Z'Z)^{-1}Z'X}, the bread matrix for 2SLS
}
