% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretest.R
\name{pretest}
\alias{pretest}
\title{Prestest estimator}
\usage{
pretest(formula, alpha = 0.05)
}
\arguments{
\item{formula}{A formula describing the model to be fitted.}

\item{alpha}{The significant level. (default = \code{0.05})}
}
\value{
\code{pretest} returns an object of class "pretest", which is a list containing the following components:
   \item{\code{coefficients}}{The estimate of the coefficients in the outcome model.}
   \item{\code{vcov}}{The estimated covariance matrix of coefficients.}
   \item{\code{Hausman.stat}}{The Hausman test statistic used to test the validity of control function.}
   \item{\code{p.value}}{The p-value of Hausman test.}
   \item{\code{cf.check}}{the indicator that the control function is valid.}
}
\description{
This function implements the pretest estimator by comparing the control function and the TSLS estimators.
}
\details{
For example, the formula \code{Y ~ D + I(D^2)+X|Z+I(Z^2)+X} describes the model where
\eqn{Y = \alpha_0 + D\beta_1 + D^2\beta_2 + X\phi + u}
and
\eqn{D = \gamma_0 + Z\gamma_1 + Z^2\gamma_2+X\psi + v}.
Here, the outcome is \code{Y}, the endogenous variables are \code{D} and \code{I(D^2)}, the baseline covariates are \code{X}, and the instrument variables are \code{Z}. The formula environment follows
the formula environment in the ivreg function in the AER package. The linear term of the endogenous variable, for example, \code{D}, must be included in must be included in the formula for the outcome model.
}
\examples{
Y <- mroz[,"lwage"]
D <- mroz[,"educ"]
Z <- as.matrix(mroz[,c("motheduc","fatheduc","huseduc")])
X <- as.matrix(mroz[,c("exper","expersq","age")])
pretest.model <- pretest(Y~D+I(D^2)+X|Z+I(Z^2)+X)
summary(pretest.model)

}
\references{
{
Guo, Z. and D. S. Small (2016), Control function instrumental variable estimation of nonlinear causal effect models, \emph{The Journal of Machine Learning Research} 17(1), 3448–3482. \cr
}
}
