% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endo-test.R
\name{print.endotest}
\alias{print.endotest}
\title{Summary of endotest}
\usage{
\method{print}{endotest}(x, ...)
}
\arguments{
\item{object}{endotest object}
}
\value{

}
\description{
Summary of endotest
}
\keyword{internal}
