% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{print.pretest}
\alias{print.pretest}
\title{Summary of pretest}
\usage{
\method{print}{pretest}(x, ...)
}
\arguments{
\item{object}{pretest object}
}
\value{

}
\description{
Summary of pretest
}
\keyword{internal}
