% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-functions.R
\name{rjive.fit}
\alias{rjive.fit}
\title{Fit RJIVE estimator given a penalty}
\usage{
rjive.fit(y, X, Z, penalty, n.endogenous)
}
\arguments{
\item{y}{An nx1 vector of outcomes}

\item{X}{An nxk matrix of regressors from the structural equation. The endogenous variables must be listed first.}

\item{Z}{An nxp matrix of regressors from the first-stage equation.}

\item{penalty}{A constant or vector > 0 to be used as Ridge penalty. Must be a vector for multiple endogenous variables. Corresponds to \gamma in the original notation.}

\item{n.endogenous}{The number of endogenous variables in X.}
}
\description{
Fits the RJIVE estimator.
}
