% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Searching-Sampling.R
\name{Searching.CI}
\alias{Searching.CI}
\title{Searching method}
\usage{
Searching.CI(
  ITT_Y,
  ITT_D,
  SigmaSqD,
  SigmaSqY,
  SigmaYD,
  InitiSet,
  WUMat,
  alpha = 0.05,
  beta.grid = NULL,
  bootstrap = TRUE
)
}
\arguments{
\item{ITT_Y}{a numeric, non-missing vector estimating each instrument's effect on the outcome}

\item{ITT_D}{a numeric, non-missing vector estimating each instrument's effect on the treatment}

\item{SigmaSqD}{a numeric, non-missing, positive scalar value estimating the error variance of Y in the reduced-form model of Y}

\item{SigmaSqY}{a numeric, non-missing, positive scalar value estimating the error variance of D in the reduced-form model of D}

\item{SigmaYD}{a numeric, non-missing, scalar value estimating the covariance of the error terms in the reduced-form models of Y and D}

\item{InitiSet}{a set of pre-selected IVs (for majority rule: it is the set of relevant IVs; for plurality rule: it is a set of weakly violated IVs)}

\item{WUMat}{a numeric matrix denoting WU where U is the precision matrix of W and W is the instrument-covariate matrix (Z, X).}

\item{alpha}{a numeric scalar value between 0 and 1 indicating the significance level for the confidence interval, with default 0.05.}

\item{beta.grid}{a numeric vector denoting candidates for betahat.}

\item{bootstrap}{a boolean scalar indicating to implement bootstrap, with default TRUE.}
}
\value{
\item{\code{CI.search}}{a two dimensional numeric vector denoting the 1-alpha confidence intervals for betaHat with lower and upper endpoints constructed by Searching method.}
\item{\code{rule}}{a boolean scalar denoting whether the identification condition is satisfied or not}
\item{\code{valid.grid}}{a numeric vector denoting the candidates satisfying certain thresholding condition}
}
\description{
CI construction by Searching method
}
