% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchingSampling.R
\name{SearchingSampling}
\alias{SearchingSampling}
\title{SearchingSampling}
\usage{
SearchingSampling(
  Y,
  D,
  Z,
  X = NULL,
  intercept = TRUE,
  method = c("OLS", "DeLasso", "Fast.DeLasso"),
  robust = TRUE,
  Sampling = TRUE,
  alpha = 0.05,
  CI.init = NULL,
  a = 0.6,
  rho = NULL,
  M = 1000,
  prop = 0.1,
  filtering = TRUE
)
}
\arguments{
\item{Y}{A vector of outcomes.}

\item{D}{A continuous vector of endogenous variables.}

\item{Z}{A matrix of instruments.}

\item{X}{A matrix of exogenous covariates.}

\item{intercept}{Should the intercept be included? Default is \code{TRUE} and if so, you do not need to add a column of 1s in X.}

\item{robust}{If \code{TRUE}, fit the model in heteroscedastic way (default=\code{TRUE})}

\item{Sampling}{If \code{TRUE}, use the proposed sampling method; else use the proposed searching method. (default=\code{TRUE})}

\item{alpha}{Significance level (default=0.05)}

\item{CI.init}{Initial interval for beta. If \code{NULL}, it will be generated automatically. (default=\code{NULL})}

\item{a}{Grid size for constructing beta grids (default=0.6)}

\item{rho}{Initial value constructed for sampling method (default=\code{NULL}). It works only for Sampling method.}

\item{M}{Sampling times (default = 1000). It works only for Sampling method.}

\item{prop}{Proportion of intervals kept when sampling (default=0.1). It works only for Sampling method.}

\item{filtering}{Filtering sampling or not (default=\code{TRUE}). It works only for Sampling method.}
}
\value{
\item{ci}{a two dimensional numeric vector denoting the 1-alpha confidence intervals for betaHat with lower and upper endpoints.}
\item{check}{True or False indicating whether the plurality rule  is satisfied or not emprically.}
\item{VHat}{a numeric vector denoting the set of valid and relevant IVs.}
\item{SHat}{a numeric vector denoting the set of relevant IVs.}
}
\description{
Conduct Searching-Sampling method, which can construct uniformly valid confidence intervals for the causal effect, which are robust to the mistakes in separating valid and invalid instruments.
}
\examples{
\dontrun{
Y <- mroz[,"lwage"]
D <- mroz[,"educ"]
Z <- as.matrix(mroz[,c("motheduc","fatheduc","huseduc","exper","expersq")])
X <- mroz[,"age"]
Searching.model <- SearchingSampling(Y,D,Z,X, Sampling = FALSE)
summary(Searching.model)
SS.model <- SearchingSampling(Y,D,Z,X)
summary(SS.model)

}
}
