% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpotIV.R
\name{SpotIV}
\alias{SpotIV}
\title{SpotIV method for causal inference in nonlinear outcome models}
\usage{
SpotIV(
  Y,
  D,
  Z,
  X = NULL,
  bs.Niter = 40,
  M = 2,
  M.est = TRUE,
  invalid = TRUE,
  intercept = TRUE,
  d1,
  d2,
  w0,
  bw = NULL,
  parallel = FALSE
)
}
\arguments{
\item{Y}{A numeric vector of outcomes.}

\item{D}{A continuous vector of endogenous variables.}

\item{Z}{A matrix of instruments.}

\item{X}{A matrix of exogenous covariates.}

\item{bs.Niter}{The number of bootstrap resampling for computing the confidence interval.}

\item{M}{The dimension of indices in the outcome model, from 1 to 3. Default is 2.}

\item{M.est}{True or False, whether estimate M based on BIC. (default = TRUE)}

\item{invalid}{If TRUE, the method is robust to the presence of possibly invalid IVs; If FALSE, the method assumes all IVs to be valid. (default = TRUE)}

\item{intercept}{Should the intercept be included? Default is \code{TRUE} and if so, you do not need to add a column of 1s in X.}

\item{d1}{a treatment value for computing CATE(d1,d2|w0).}

\item{d2}{a treatment value for computing CATE(d1,d2|w0).}

\item{w0}{a value of measured covariates and instruments for computing CATE(d1,d2|w0).}

\item{bw}{a (M+1) by 1 vector bandwidth specification. Default is NULL and the bandwidth is chosen by rule of thumb.}

\item{parallel}{True or False indicating whether to use parallel computing (maybe useless on Windows). (default = FALSE)}
}
\value{
\code{ProbitControl} returns an object of class "SpotIV".
    An object class "SpotIV" is a list containing the following components:
    \item{\code{betaHat}}{a numeric scalar denoting the estimate of beta.}
    \item{\code{cateHat}}{a numeric scalar denoting the estimate of CATE(d1,d2|w0).}
    \item{\code{cate.sdHat}}{a numeric scalar denoting the estimated standard deviation of cateHat.}
    \item{\code{SHat}}{a numeric vector denoting the set of relevant IVs.}
    \item{\code{VHat}}{a numeric vector denoting the set of relevant and valid IVs.}
    \item{\code{Maj.pass}}{True or False indicating whether the majority rule test is passed or not.}
}
\description{
Perform causal inference in semi-parametric outcome model with possibly invalid IVs under the majority rule using SpotIV method.
}
\examples{
\dontrun{
Y <- mroz[,"lwage"]
D <- mroz[,"educ"]
Z <- as.matrix(mroz[,c("motheduc","fatheduc","huseduc","exper","expersq")])
X <- mroz[,"age"]
Y0 <- as.numeric((Y>median(Y)))
d2 = median(D); d1 = d2+1;
w0 = apply(cbind(Z,X)[which(D == d2),], 2, mean)
SpotIV.model <- SpotIV(Y0,D,Z[,-5],X,d1 = d1,d2 = d2,w0 = w0[-5])
summary(SpotIV.model)
}



}
