% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-endo.test.R
\name{summary.endotest}
\alias{summary.endotest}
\title{Summary of endotest}
\usage{
\method{summary}{endotest}(object, ...)
}
\value{
No return value, called for summary.
}
\description{
Summary function for endo.test
}
\keyword{internal}
