% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretest.R
\name{summary.pretest}
\alias{summary.pretest}
\title{Summary of pretest}
\usage{
\method{summary}{pretest}(object, ...)
}
\arguments{
\item{object}{pretest object}
}
\value{

}
\description{
Summary of pretest
}
\keyword{internal}
