% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchingSampling.R
\name{summary.SS}
\alias{summary.SS}
\title{Summary of SS}
\usage{
\method{summary}{SS}(object, ...)
}
\value{
No return value, called for summary.
}
\description{
Summary function for SearchingSampling
}
\keyword{internal}
