% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-TSHT.R
\name{summary.TSHT}
\alias{summary.TSHT}
\title{Summary of TSHT}
\usage{
\method{summary}{TSHT}(object, ...)
}
\value{
No return value, called for summary.
}
\description{
Summary function for TSHT
}
\keyword{internal}
