% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Searching-Sampling.R
\name{TSHT.Initial}
\alias{TSHT.Initial}
\title{Two Stage Hard Thresholding for initial estimator.}
\usage{
TSHT.Initial(
  ITT_Y,
  ITT_D,
  WUMat,
  SigmaSqY,
  SigmaSqD,
  SigmaYD,
  covW,
  alpha = 0.05,
  bootstrap = FALSE,
  tuning = 2.01,
  max_clique = FALSE
)
}
\arguments{
\item{ITT_Y}{a numeric, non-missing vector estimating each instrument's effect on the outcome}

\item{ITT_D}{a numeric, non-missing vector estimating each instrument's effect on the treatment}

\item{WUMat}{an n by pz numeric, non-missing matrix that computes the precision of W
(ex1 t(WUMat) %*% WUMat / n = (W^T W/n)^(-1))  (ex2 U = (W^T W/n)^(-1))}

\item{SigmaSqY}{a numeric, non-missing, positive scalar value estimating the error variance of Y in the reduced-form model of Y}

\item{SigmaSqD}{a numeric, non-missing, positive scalar value estimating the error variance of D in the reduced-form model of D}

\item{SigmaYD}{a numeric, non-missing, scalar value estimating the covariance of the error terms in the reduced-form models of Y and D}

\item{alpha}{a numeric scalar value between 0 and 1 indicating the significance level for the confidence interval, with default 0.05.}

\item{bootstrap}{a boolean scalar indicating to implement bootstrap, with default TRUE.}

\item{tuning}{a numeric scalar value tuning parameter for TSHT greater than 2 (default = 2.01)}

\item{max_clique}{an option to replace the majority and plurality voting procedures with finding maximal clique in the IV voting matrix, with default FALSE.}
}
\value{
\item{\code{VHat}}{a numeric vector denoting the set of valid and relevant IVs.}
    \item{\code{SHat}}{a numeric vector denoting the set of relevant IVs.}
}
\description{
This modifies the original TSHT function and produces an initial esitmator for sampling and searching
}
