% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSHT.R
\name{TSHT.OLS}
\alias{TSHT.OLS}
\title{Two-Stage Hard Thresholding Ordinary Least Squares}
\usage{
TSHT.OLS(Y, D, W, pz, intercept = TRUE)
}
\arguments{
\item{Y}{continuous and non-missing, n by 1 numeric outcome vector.}

\item{D}{continuous or discrete, non-missing, n by 1 numeric treatment vector.}

\item{W}{continuous or discrete, non-missing, n by (p_z + p_x) numeric instrument-covariate matrix.}

\item{pz}{a numeric scalar denoting the number of instrument variables.}

\item{intercept}{a boolean scalar indicating to include the intercept or not, with default TRUE.}
}
\value{
\item{\code{ITT_Y}}{a p_z by 1 numeric vector denoting the estimated coefficients of instruments in the treatment model.}
    \item{\code{ITT_D}}{a p_z by 1 numeric vector denoting the estimated coefficients of instruments in the outcome model.}
    \item{\code{WUMat}}{a numeric matrix denoting WU where U is the precision matrix of W and W is the instrument-covariate matrix.}
    \item{\code{SigmaSqY}}{a numeric scalar denoting the consistent estimator of the noise level in the outcome model.}
    \item{\code{SigmaSqY}}{a numeric scalar denoting the consistent estimator of the noise level in the treatment model.}
    \item{\code{SigmaYD}}{a numeric scalar denoting the consistent estimator of the covariance between the error term in the treatment model and the error term in the outcome model.}
}
\description{
This function provides the estimates of the necessary inputs of TSHT method in a low-dimensional setting using Ordinary Least Squares.
}
