% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSHT.R
\name{TSHT}
\alias{TSHT}
\title{Two-Stage Hard Thresholding}
\usage{
TSHT(
  Y,
  D,
  Z,
  X,
  intercept = TRUE,
  alpha = 0.05,
  method = "OLS",
  voting = "MaxClique",
  robust = FALSE
)
}
\arguments{
\item{Y}{A continuous vector of outcomes.}

\item{D}{A continuous vector of endogenous variables.}

\item{Z}{A matrix of instruments.}

\item{X}{A matrix of exogenous covariates.}

\item{intercept}{Should the intercept be included? Default is \code{TRUE} and if so, you do not need to add a column of 1s in X.}

\item{alpha}{The significance level for the confidence interval. (default = 0.05)}

\item{method}{The method which will be used to estimate the reduced form parameters in TSHT. "OLS" stands for ordinary least square and "DeLasso" stands for the debiased Lasso estimator. (default = "OLS")}

\item{voting}{The voting option used to estimate valid IVs. 'MP' stands for majority and plurality voting, 'MaxClique' stands for finding maximal clique in the IV voting matrix, and 'Conservative' stands for conservative voting procedure. Conservative voting is used to get an initial estimator of valid IVs in the Searching-Sampling method. (default= 'MaxClique').}

\item{robust}{If \code{TRUE}, the method is robust to heteroskedasticity errors. If \code{FALSE}, the method assumes homoskedasticity errors.  When robust = \code{TRUE}, only 'OLS' can be input to \code{method}. (default = \code{FALSE})}
}
\value{
\code{TSHT} returns an object of class "TSHT".
    An object class "TSHT" is a list containing the following components:
    \item{\code{betaHat}}{The estimate of treatment effect.}
    \item{\code{beta.sdHat}}{The estimated standard deviation of \code{betaHat}.}
    \item{\code{ci}}{The 1-alpha confidence intervals for \code{betaHat}.}
    \item{\code{SHat}}{The set of relevant IVs.}
    \item{\code{VHat}}{The set of valid and relevant IVs.}
    \item{\code{voting.mat}}{The voting matrix on whether the elements of each \code{SHat} are valid or not.}
    \item{\code{check}}{Whether the majority rule test is passed or not.}
    \item{\code{beta.clique}}{The estimates of treatment effect from each maximum clique. Only returns when \code{voting} is \code{'MaxClique'}.}
    \item{\code{beta.sd.clique}}{The estimated standard deviation of \code{betaHat} of each maximum clique. Only returns when \code{voting} is \code{'MaxClique'}}
    \item{\code{CI.clique}}{The 1-alpha confidence intervals for \code{betaHat} of each maximum clique. Only returns when \code{voting} is \code{'MaxClique'}}
    \item{\code{max.clique}}{The maximum cliques of voted as valid IVs. Only returns when \code{voting} is \code{'MaxClique'}}
}
\description{
Perform Two-Stage Hard Thresholding method, which provides the robust inference of the treatment effect in the presence of invalid instrumental variables in both low-dimensional and high-dimensional settings.
}
\examples{
\dontrun{
Y <- mroz[,"lwage"]
D <- mroz[,"educ"]
Z <- as.matrix(mroz[,c("motheduc","fatheduc","huseduc","exper","expersq")])
X <- mroz[,"age"]
TSHT.model <- TSHT(Y=Y,D=D,Z=Z,X=X)
summary(TSHT.model)
}


}
