% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSHT.R
\name{TSHT.VHat}
\alias{TSHT.VHat}
\title{Two-Stage Hard Thresholding Instrumental Variable Selection}
\usage{
TSHT.VHat(
  ITT_Y,
  ITT_D,
  WUMat,
  SigmaSqY,
  SigmaSqD,
  SigmaYD,
  bootstrap = FALSE,
  tuning = 2.01,
  max_clique
)
}
\arguments{
\item{ITT_Y}{a p_z by 1 numeric vector denoting the estimated coefficients of instruments in the treatment model.}

\item{ITT_D}{a p_z by 1 numeric vector denoting the estimated coefficients of instruments in the treatment model.}

\item{WUMat}{a numeric matrix denoting WU where U is the precision matrix of W and W is the instrument-covariate matrix (Z, X).}

\item{SigmaSqY}{a numeric scalar denoting the consistent estimator of the noise level in the outcome model.}

\item{SigmaSqD}{a numeric scalar denoting the consistent estimator of the noise level in the treatment model.}

\item{SigmaYD}{a numeric scalar denoting the consistent estimator of the covariance between the error term in the treatment model and the error term in the outcome model.}

\item{bootstrap}{a logical value, default by FALSE(What is this for in TSHT.Initial?).}

\item{tuning}{a numeric scalar value tuning parameter for TSHT greater 2, with default 2.01.}

\item{max_clique}{an option to replace the majority and plurality voting procedures with finding maximal clique in the IV voting matrix.}
}
\value{
\item{\code{VHat}}{a numeric vector denoting the set of valid and relevant IVs.}
    \item{\code{SHat}}{a numeric vector denoting the set of relevant IVs.}
    \item{\code{max.clique}}{a numeric list denoting the maximum cliques of valid and relevant IVs. Only return when \code{max_clique} is \code{TRUE}.}
}
\description{
Implementation of the Two-Stage Hard Thresholding with Voting procedures.This function only takes the necessary inputs of the functions \code{\link{TSHT.OLS}} or \code{\link{TSHT.DeLasso}}. Any other methods to compute the necessary inputs can be adopted by the users according to their preferences.
}
