% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-functions.R
\name{tsls.fit}
\alias{tsls.fit}
\title{Fit 2SLS estimator}
\usage{
tsls.fit(y, X, Z)
}
\arguments{
\item{y}{An nx1 vector of outcomes}

\item{X}{An nxk matrix of regressors from the structural equation}

\item{Z}{An nxp matrix of regressors from the first-stage equation}
}
\description{
Fits the 2SLS estimator given the outcomes, instruments and structural regressors.
}
