% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asymptotic-var-cov.R
\name{tsls.vcov}
\alias{tsls.vcov}
\title{Calculate 2SLS asymptotic var-cov matrix}
\usage{
tsls.vcov(X, Z, e.hat, type = "hc1")
}
\arguments{
\item{X}{An nxk matrix of regressors from the structural equation}

\item{Z}{An nxp matrix of regressors from the first-stage equation}

\item{e.hat}{An nx1 vector of residuals from the structural equation}

\item{type}{The type of var-cov matrix to calculate. Defaults to HC1. Currently supports homoskedastic SEs, HC0 & HC1}
}
\description{
Calculates the 2SLS var-cov matrix given asymptotic theory
}
