% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf}
\alias{cf}
\title{Control-Function}
\usage{
cf(formula, d1 = NULL, d2 = NULL)
}
\arguments{
\item{formula}{a formula describing the model to be fitted. For example, the formula \code{Y ~ D + I(D^2)+X|Z+X} describes the model where
\eqn{Y = \alpha_0 + D\beta_1 + D^2\beta_2 + X\phi + u}
and
\eqn{D = \gamma_0 + Z\gamma_1 + X\psi + v}.
Here, the outcome is \code{Y}, the endogenous variables are \code{D} and \code{I(D^2)}, the exogenous covariates are \code{X}, and the instrument variables are \code{Z}. The formula environment follows
the formula environment in the ivreg function in the AER package. The linear term of the endogenous variable, for example, \code{D}, must be included in the first of the right side of the formula.}

\item{d1}{a treatment value for calculating causal effect from \code{d2} to \code{d1}.}

\item{d2}{a treatment value for calculating causal effect from \code{d2} to \code{d1}.}
}
\value{
\code{cf} returns an object of class "cf".
    An object class "cf" is a list containing the following components:
   \item{\code{coefficients}}{The estimate of the coefficients in the outcome model.}
   \item{\code{vcov}}{The estimated covariance matrix of coefficients.}
   \item{\code{CausalEffect}}{The causal effect of increasing endogenous variable from \code{d2} to \code{d1}. If one of \code{d2} or \code{d1} is \code{NULL}, it returns \code{NULL}}
}
\description{
Implement the control function method for estimation and inference of nonlinear treatment effects as developed in Guo and Small (2016).
}
\examples{
\dontrun{
Y <- mroz[,"lwage"]
D <- mroz[,"educ"]
Z <- as.matrix(mroz[,c("motheduc","fatheduc","huseduc")])
X <- as.matrix(mroz[,c("exper","expersq","age")])
cf.model <- cf(Y~D+I(D^2)+X|Z+I(Z^2)+X,d1 = c(median(D)+1,(median(D)+1)^2),d2 = c(median(D),median(D)^2))
summary(cf.model)
}


}
