% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf}
\alias{cf}
\title{Control-Function}
\usage{
cf(outcome.formula, treatment.formula)
}
\arguments{
\item{outcome.formula}{formula specification(s) of the regression relationship between outcome, covariates and the endogeneity variables, such as \code{Y ~ X + D + g_2(D) + ... + g_k(D)}.}

\item{treatment.formula}{formula specification(s) of the regression relationship between endogeneity variables, covariates and the instrumental variables, such as \code{D ~ X + Z + h_2(Z) + ... + h_k(Z)}.}
}
\value{
\item{\code{coefficients}}{scalar numeric value: the estimate of the treatment effect }
   \item{\code{vcov}}{numeric matrix: estimated covariance matrix of coefficients}
}
\description{
This function implements the control function method for estimation and inference of nonlinear treatment effects as developed in Guo and Small (2016)
}
\examples{

### Generate the data
library(MASS)
library(RobustIV)
n <- 10000
mu <- rep(0,2); V <- cbind(c(1,0.5),c(0.5,1))
X <- rnorm(n); Z <- rnorm(n)
err <- mvrnorm(n,mu=mu,Sigma = V)
u1 <- err[,1]; v2 <- err[,2]
D <- 1+X/8+Z/3+Z^2/8+v2
Y <- 1+X+10*D +10*D^2+u1

### Implement the control function method
cf(Y~X+D+I(D^2),D~X+Z+I(Z^2))


}
