% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endo-test.R
\name{endo.test}
\alias{endo.test}
\title{Endogeneity test in high dimensions}
\usage{
endo.test(
  Y,
  D,
  Z,
  X,
  intercept = TRUE,
  alpha = 0.05,
  method = c("Fast.DeLasso", "DeLasso", "OLS"),
  invalid = FALSE,
  voting = "MaxClique"
)
}
\arguments{
\item{Y}{The outcome observation, a vector of length \eqn{n}.}

\item{D}{The treatment observation, a vector of length \eqn{n}.}

\item{Z}{The instrument observation of dimension \eqn{n \times p_z}.}

\item{X}{The covariates observation of dimension \eqn{n \times p_x}.}

\item{intercept}{Whether the intercept is included. (default = \code{TRUE})}

\item{alpha}{The significance level for the confidence interval. (default = 0.05)}

\item{method}{The method used to estimate the reduced form parameters. "OLS" stands for ordinary least squares, "DeLasso" stands for the debiased Lasso estimator, and "Fast.DeLasso" stands for the debiased Lasso estimator with fast algorithm. (default = "Fast.DeLasso")}

\item{invalid}{If \code{TRUE}, the method is robust to the presence of possibly invalid IVs; If \code{FALSE}, the method assumes all IVs to be valid. (default = \code{FALSE})}

\item{voting}{The voting option used to estimate valid IVs. 'MP' stnads for majority and plurality voting, 'MaxClique' stands for maximum clique in the IV voting matrix. (default = 'MaxClique')}
}
\value{
\code{endo.test} returns an object of class "endotest".
    An object class "endotest" is a list containing the following components:
   \item{\code{Q}}{Ttest statistic.}
   \item{\code{Sigma12}}{Estimated covaraince.}
   \item{\code{VHat}}{The estimated set of relevant and vaild IVs.}
   \item{\code{alpha}}{The significance level.}
}
\description{
Conduct the endogeneity test with with high dimensional and possibly invalid instrumental variables.
}
\examples{
\dontrun{
n = 500; L = 600; s = 3; k = 10; px = 10;
alpha = c(rep(3,s),rep(0,L-s)); beta = 1; gamma = c(rep(1,k),rep(0,L-k))
phi<-(1/px)*seq(1,px)+0.5; psi<-(1/px)*seq(1,px)+1
epsilonSigma = matrix(c(1,0.8,0.8,1),2,2)
Z = matrix(rnorm(n*L),n,L)
X = matrix(rnorm(n*px),n,px)
epsilon = MASS::mvrnorm(n,rep(0,2),epsilonSigma)
D =  0.5 + Z \%*\% gamma + X \%*\% psi + epsilon[,1]
Y = -0.5 + Z \%*\% alpha + D * beta + X \%*\% phi + epsilon[,2]
TSHT(Y,D,Z,X, method = "DeLasso")
endo.test.model <- endo.test(Y,D,Z,X)
summary(endo.test.model)
}

}
\references{
{
Guo, Z., Kang, H., Tony Cai, T. and Small, D.S. (2018), Testing endogeneity with high dimensional covariates, \emph{Journal of Econometrics}, Elsevier, vol. 207(1), pages 175-187. \cr
}
}
