% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchingSampling.R
\name{SearchingSampling}
\alias{SearchingSampling}
\title{Searching-Sampling}
\usage{
SearchingSampling(
  Y,
  D,
  Z,
  X = NULL,
  intercept = TRUE,
  method = c("OLS", "DeLasso", "Fast.DeLasso"),
  robust = FALSE,
  Sampling = TRUE,
  alpha = 0.05,
  CI.init = NULL,
  a = 0.6,
  rho = NULL,
  M = 1000,
  prop = 0.1,
  filtering = TRUE
)
}
\arguments{
\item{Y}{The outcome observation, a vector of length \eqn{n}.}

\item{D}{The treatment observation, a vector of length \eqn{n}.}

\item{Z}{The instrument observation of dimension \eqn{n \times p_z}.}

\item{X}{The covariates observation of dimension \eqn{n \times p_x}.}

\item{intercept}{Whether the intercept is included. (default = \code{TRUE})}

\item{method}{The method used to estimate the reduced form parameters. "OLS" stands for ordinary least squares, "DeLasso" stands for the debiased Lasso estimator, and "Fast.DeLasso" stands for the debiased Lasso estimator with fast algorithm. (default = \code{"OLS"})}

\item{robust}{If \code{TRUE}, the method is robust to heteroskedasticity errors. If \code{FALSE}, the method assumes homoskedasticity errors. When \code{robust = TRUE}, only \code{’OLS’} can be input to \code{method}. (default = \code{FALSE})}

\item{Sampling}{If \code{TRUE}, use the proposed sampling method; else use the proposed searching method. (default=\code{TRUE})}

\item{alpha}{Significance level (default=0.05)}

\item{CI.init}{Initial interval for beta. If \code{NULL}, it will be generated automatically. (default=\code{NULL})}

\item{a}{Grid size for constructing beta grids (default=0.6)}

\item{rho}{Initial value constructed for sampling method (default=\code{NULL}). It works only for Sampling method.}

\item{M}{Sampling times (default = 1000). It works only for Sampling method.}

\item{prop}{Proportion of intervals kept when sampling (default=0.1). It works only for Sampling method.}

\item{filtering}{Filtering sampling or not (default=\code{TRUE}). It works only for Sampling method.}
}
\value{
\item{ci}{1-alpha confidence interval for beta.}
\item{check}{Indicator for whether the plurality rul is satisfied.}
\item{SHat}{The set of relevant IVs.}
\item{VHat}{The initial set of relevant and valid IVs.}
}
\description{
Construct Searching and Sampling confidence intervals for the causal effect, which are robust to the mistakes in separating valid and invalid instruments.
}
\examples{
\dontrun{
Y <- mroz[,"lwage"]
D <- mroz[,"educ"]
Z <- as.matrix(mroz[,c("motheduc","fatheduc","huseduc","exper","expersq")])
X <- mroz[,"age"]
Searching.model <- SearchingSampling(Y,D,Z,X, Sampling = FALSE)
summary(Searching.model)
SS.model <- SearchingSampling(Y,D,Z,X)
summary(SS.model)

}
}
\references{
{
Guo, Z. (2021), Causal Inference with Invalid Instruments: Post-selection Problems and A Solution Using Searching and Sampling, Preprint \emph{arXiv:2104.06911}. \cr
}
}
