% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpotIV.R
\name{SpotIV}
\alias{SpotIV}
\title{SpotIV method for causal inference in semi-parametric outcome model}
\usage{
SpotIV(
  Y,
  D,
  Z,
  X = NULL,
  invalid = TRUE,
  intercept = TRUE,
  d1,
  d2,
  w0,
  M.est = TRUE,
  bs.Niter = 40,
  M = 2,
  bw = NULL
)
}
\arguments{
\item{Y}{The outcome observation, a vector of length \eqn{n}.}

\item{D}{The treatment observation, a vector of length \eqn{n}.}

\item{Z}{The instrument observation of dimension \eqn{n \times p_z}.}

\item{X}{The covariates observation of dimension \eqn{n \times p_x}.}

\item{invalid}{If TRUE, the method is robust to the presence of possibly invalid IVs; If FALSE, the method assumes all IVs to be valid. (default = TRUE)}

\item{intercept}{Should the intercept be included? Default is \code{TRUE} and if so, you do not need to add a column of 1s in X.}

\item{d1}{A treatment value for computing CATE(d1,d2|w0).}

\item{d2}{A treatment value for computing CATE(d1,d2|w0).}

\item{w0}{A value of measured covariates and instruments for computing CATE(d1,d2|w0).}

\item{M.est}{Whether estimate M based on BIC. (default = TRUE)}

\item{bs.Niter}{The number of bootstrap resampling for computing the confidence interval.}

\item{M}{The dimension of indices in the outcome model, from 1 to 3. Default is 2.}

\item{bw}{A (M+1) by 1 vector bandwidth specification. Default is NULL and the bandwidth is chosen by rule of thumb.}
}
\value{
\code{SpotIV} returns an object of class "SpotIV".
    An object class "SpotIV" is a list containing the following components:
    \item{\code{betaHat}}{a numeric scalar denoting the estimate of beta.}
    \item{\code{cateHat}}{a numeric scalar denoting the estimate of CATE(d1,d2|w0).}
    \item{\code{cate.sdHat}}{a numeric scalar denoting the estimated standard error of cateHat.}
    \item{\code{SHat}}{a numeric vector denoting the set of relevant IVs.}
    \item{\code{VHat}}{a numeric vector denoting the set of relevant and valid IVs.}
    \item{\code{Maj.pass}}{True or False indicating whether the majority rule test is passed or not.}
}
\description{
Perform causal inference in the semi-parametric outcome model with possibly invalid IVs.
}
\examples{
\dontrun{
Y <- mroz[,"lwage"]
D <- mroz[,"educ"]
Z <- as.matrix(mroz[,c("motheduc","fatheduc","huseduc","exper","expersq")])
X <- mroz[,"age"]
Y0 <- as.numeric((Y>median(Y)))
d2 = median(D); d1 = d2+1;
w0 = apply(cbind(Z,X)[which(D == d2),], 2, mean)
SpotIV.model <- SpotIV(Y0,D,Z[,-5],X,d1 = d1,d2 = d2,w0 = w0[-5])
summary(SpotIV.model)
}

}
\references{
{
Li, S., Guo, Z. (2020), Causal Inference for Nonlinear Outcome Models with Possibly Invalid Instrumental Variables, Preprint \emph{arXiv:2010.09922}.\cr
}
}
