% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endo-test.R
\name{summary.endotest}
\alias{summary.endotest}
\title{Summary of endotest}
\usage{
\method{summary}{endotest}(object, ...)
}
\arguments{
\item{object}{endotest object}
}
\description{
Summary of endotest
}
\keyword{internal}
