% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSHT.R
\name{TSHT}
\alias{TSHT}
\title{Two-Stage Hard Thresholding}
\usage{
TSHT(
  Y,
  D,
  Z,
  X,
  intercept = TRUE,
  alpha = 0.05,
  method = c("OLS", "DeLasso", "Fast.DeLasso"),
  voting = "MaxClique",
  robust = FALSE
)
}
\arguments{
\item{Y}{The outcome observation, a vector of length \eqn{n}.}

\item{D}{The treatment observation, a vector of length \eqn{n}.}

\item{Z}{The instrument observation of dimension \eqn{n \times p_z}.}

\item{X}{The covariates observation of dimension \eqn{n \times p_x}.}

\item{intercept}{Whether the intercept is included. (default = \code{TRUE})}

\item{alpha}{The significance level for the confidence interval. (default = 0.05)}

\item{method}{The method used to estimate the reduced form parameters. \code{"OLS"} stands for ordinary least squares, \code{"DeLasso"} stands for the debiased Lasso estimator, and \code{"Fast.DeLasso"} stands for the debiased Lasso estimator with fast algorithm. (default = \code{"OLS"})}

\item{voting}{The voting option used to estimate valid IVs. 'MP' stands for majority and plurality voting, 'MaxClique' stands for finding maximal clique in the IV voting matrix, and 'Conservative' stands for conservative voting procedure. Conservative voting is used to get an initial estimator of valid IVs in the Searching-Sampling method. (default= 'MaxClique').}

\item{robust}{If \code{TRUE}, the method is robust to heteroskedasticity errors. If \code{FALSE}, the method assumes homoskedasticity errors. When \code{robust = TRUE}, only \code{’OLS’} can be input to \code{method}. (default = \code{FALSE})}
}
\value{
\code{TSHT} returns an object of class "TSHT".
    An object class "TSHT" is a list containing the following components:
    \item{\code{betaHat}}{The estimate of treatment effect.}
    \item{\code{beta.sdHat}}{The estimated standard error of \code{betaHat}.}
    \item{\code{ci}}{The 1-alpha confidence interval for \code{beta}.}
    \item{\code{SHat}}{The set of relevant IVs.}
    \item{\code{VHat}}{The set of relevant and valid IVs.}
    \item{\code{voting.mat}}{The voting matrix on whether the elements of each \code{SHat} are valid or not.}
    \item{\code{check}}{Whether the majority rule test is passed or not.}
    \item{\code{beta.clique}}{The estimates of treatment effect from each maximum clique. Only returns when \code{voting} is \code{'MaxClique'}.}
    \item{\code{beta.sd.clique}}{The estimated standard deviation of \code{betaHat} of each maximum clique. Only returns when \code{voting} is \code{'MaxClique'}}
    \item{\code{CI.clique}}{The 1-alpha confidence interval for \code{beta} of each maximum clique. Only returns when \code{voting} is \code{'MaxClique'}}
    \item{\code{max.clique}}{The maximum cliques voted as valid IVs. Only returns when \code{voting} is \code{'MaxClique'}}
}
\description{
Perform Two-Stage Hard Thresholding method, which provides the robust inference of the treatment effect in the presence of invalid instrumental variables in both low-dimensional and high-dimensional settings.
}
\examples{
\dontrun{
Y <- mroz[,"lwage"]
D <- mroz[,"educ"]
Z <- as.matrix(mroz[,c("motheduc","fatheduc","huseduc","exper","expersq")])
X <- mroz[,"age"]
TSHT.model <- TSHT(Y=Y,D=D,Z=Z,X=X)
summary(TSHT.model)
}
}
\references{
{
Guo, Z., Kang, H., Tony Cai, T. and Small, D.S. (2018), Confidence intervals for causal effects with invalid instruments by using two-stage hard thresholding with voting, \emph{J. R. Stat. Soc. B}, 80: 793-815. \cr
}
}
